Re-modularised the context menus
