+++ src/skins/vector/css/common.css
@@ -85,87 +85,6 @@ input[type=text]:focus, textarea:focus {
     border-left: 6px solid transparent;
 }
 
-.mx_ContextualMenu_background {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    opacity: 1.0;
-    z-index: 2000;
-}
-
-.mx_ContextualMenu {
-    border: solid 1px rgba(187, 187, 187, 0.5);
-    border-radius: 4px;
-    background-color: #f6f6f6;
-    color: #4a4a4a;
-    position: fixed;
-    z-index: 2001;
-    padding: 6px;
-	font-size: 14px;
-}
-
-.mx_ContextualMenu_chevron_right {
-    position: absolute;
-    right: -8px;
-    top: 10px;
-    width: 0;
-    height: 0;
-    border-top: 8px solid transparent;
-    border-left: 8px solid rgba(187, 187, 187, 0.5);
-    border-bottom: 8px solid transparent;
-}
-
-.mx_ContextualMenu_chevron_right:after{
-    content:'';
-    width: 0;
-    height: 0;
-    border-top: 7px solid transparent;
-    border-left: 7px solid #f6f6f6;
-    border-bottom: 7px solid transparent;
-    position:absolute;
-    top: -7px;
-    right: 1px;
-}
-
-.mx_ContextualMenu_chevron_left {
-    position: absolute;
-    left: -8px;
-    top: 10px;
-    width: 0;
-    height: 0;
-    border-top: 8px solid transparent;
-    border-right: 8px solid rgba(187, 187, 187, 0.5);
-    border-bottom: 8px solid transparent;
-}
-
-.mx_ContextualMenu_chevron_left:after{
-    content:'';
-    width: 0;
-    height: 0;
-    border-top: 7px solid transparent;
-    border-right: 7px solid #f6f6f6;
-    border-bottom: 7px solid transparent;
-    position:absolute;
-    top: -7px;
-    left: 1px;
-}
-
-.mx_ContextualMenu_field {
-    padding: 3px 6px 3px 6px;
-    cursor: pointer;
-}
-
-.mx_ContextualMenu_field.mx_ContextualMenu_fieldSet {
-    font-weight: bold;
-}
-
-.mx_ContextualMenu_spinner {
-    display: block;
-    margin: 0 auto;
-}
-
 .mx_Dialog_wrapper {
     position: fixed;
     z-index: 4000;
